+++ client/src/features/interview-record/InterViewSlider.tsx
